rename: style file tsx to ts
