+++ src/components/AIChat/ChatNavBar.styles.ts
