+++ src/components/AIChat/Message/Message.styles.ts
@@ -1,24 +1,5 @@
 import { css } from '@emotion/react';
 
-export const messagesContainer = css`
-  padding: 0px 20px;
-  width: 100%;
-  flex: 1;
-  overflow-y: scroll;
-
-  &::before {
-    content: '';
-    display: block;
-    margin-top: -32px;
-  }
-
-  &::after {
-    content: '';
-    display: block;
-    height: 48px;
-  }
-`;
-
 export const userContainer = css`
   padding: 32px 32px 24px 32px;
 `;
