+++ src/components/ProductDetail/ProductInfo.styles.ts
