+++ src/components/ProductDetail/TabbedContent.styles.ts
