+++ src/components/UI/Card/ContentCard.styles.ts
